fix scan_binding() handling empty data set in chart_cell.ex
Issue is `infer_types({:rows, _, _})` assumes that there is
at least 1 row so that it can `Enum.map` over the columns.

We will check in `columns_for` that the list of columns is
not empty.  If so, and the dataset is empty, then the new
implementation of `infer_types()` will return a list of
`nil`s to allow the scan to succeed but without inferred types.